Finished and improved FLAS and sample grid.

+++ Saempl/Source/BlomeFileFilterRuleViewLoudnessDecibel.cpp
@@ -69,7 +69,7 @@ void BlomeFileFilterRuleViewLoudnessDecibel::textEditorFocusLost(TextEditor& tex
 {
     // Lose focus, set compare value and refresh library
     mCompareValueEditor->giveAwayKeyboardFocus();
-    getFilterRule().setCompareValue(textEditor.getText().getIntValue());
+    getFilterRule().setCompareValue(textEditor.getText().getFloatValue());
     sampleLibrary.refresh();
 }
